# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Forum Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/answers/{comment}/{receiver}": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Добавить ответ на комментарий",
                "description": "Создает новый ответ на указанный комментарий и отправляет его указанному получателю.",
                "operationId": "4f7513f4014201cc87106a8e7e3bc072",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID комментария, на который добавляется ответ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "receiver",
                        "in": "path",
                        "description": "ID пользователя, которому отправляется ответ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Содержимое ответа",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное создание ответа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "comment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "content": {
                                                    "type": "string"
                                                },
                                                "comment_id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "receiver_id": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Неверный запрос."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нет прав на ресурс",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 403
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "У вас нет прав на этот ресурс."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ресурс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Запрашиваемый ресурс не найден."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 404
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Вы не авторизированны. Пожалуйста пройдите авторизацию и возвращайтесь."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 422
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Введены некорректные данные. Пожалуйста пересмотрите свой запрос и попробуйте снова."
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fault": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Произошла непредвиденная ошибка."
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}